Added Display All functionality for Leads.
Modified MainMenuController to add Display All Leads functionality, and
to be able to pass down the MainMenuView to other controllers. Modified
MainMenuView to add the ability to add and remove tabs.  Added
DisplayAllLeadsController and DisplayAllLeadsView, to act as the
controller and view for the Display All Leads use case, respectively.

GUI for this use case may require updating in future, but for now
functions as needed.